feat: add types to emergency provider
